Revert "Use su mode by default so the user can save the password to session"
This reverts commit af678be1e08ae23cf0cf727d3fc128a8e367948e.

Even with all that, and despite postinst setting things in su mode, we still end-up in sudo mode when building an ISO.
That basically means there's something hard-wired deep into the OS upstream base layer to make sure we're in sudo mode.
The wiser decision is to give up on this and recommend people use gksu-properties instead.

Note: libpam-systemd was fixed to allow su mode.